requirements.txt
python-dotenv
langchain-openai
streamlit